+++ workers/tests/test_pg_queue_consumer.py
@@ -12,12 +12,10 @@ import logging
 import os
 from unittest.mock import MagicMock
 
-import psycopg2
 import pytest
 from celery import shared_task
-from queue_backend.pg_queue import PgQueueClient, to_payload
+from queue_backend.pg_queue import to_payload
 from queue_backend.pg_queue.client import QueueMessage
-from queue_backend.pg_queue.connection import create_pg_connection
 from queue_backend.pg_queue.consumer import PgQueueConsumer
 
 # Registered test tasks (namespaced). apply() runs their bodies in-process.
@@ -96,27 +94,7 @@ class TestRunLoop:
 
 
 # --- Integration: full enqueue → poll → execute → ack against real PG ---
-
-
-def _test_conn():
-    os.environ.setdefault("TEST_DB_HOST", "127.0.0.1")
-    return create_pg_connection(env_prefix="TEST_DB_")
-
-
-@pytest.fixture
-def pg_client():
-    try:
-        conn = _test_conn()
-    except psycopg2.OperationalError as exc:
-        pytest.skip(f"Postgres not reachable: {exc}")
-    with conn.cursor() as cur:
-        cur.execute("SELECT to_regclass('pg_queue_message')")
-        if cur.fetchone()[0] is None:
-            conn.close()
-            pytest.skip("pg_queue migration not applied (run backend migrate)")
-    yield PgQueueClient(conn=conn)
-    conn.rollback()
-    conn.close()
+# Uses the shared ``pg_client`` fixture from conftest.py.
 
 
 class TestConsumerIntegration:
